Migrated from manual Service Locator to Koin-DI. LocalIcon id is now Int. Naming has been improved: viewmodels have been renamed, and Flow has been renamed to NavGraph in navigation. The Dispatchers.IO passed to the logics have been replaced with a new BackgroundCoroutineDispatchers.kt entity (if you need to switch, for example, for calculations on Dispatchers.Default)

+++ app/android-app/src/main/kotlin/hardcoder/dev/androidApp/ui/features/diary/tags/providers/DiaryTagIconProvider.kt
@@ -28,7 +28,7 @@ class DiaryTagIconProvider : IconResourceProvider {
 
     override fun getIcons(): List<LocalIcon> = icons
 
-    override fun getIcon(id: Long): LocalIcon = icons.first { it.id == id }
+    override fun getIcon(id: Int): LocalIcon = icons.first { it.id == id }
 
-    private fun create(id: Long, resourceId: Int) = LocalIconImpl(id, resourceId)
+    private fun create(id: Int, resourceId: Int) = LocalIconImpl(id, resourceId)
 }
