+++ app/android-app/src/main/kotlin/hardcoder/dev/androidApp/ui/features/moodTracking/moodType/providers/PredefinedMoodTypeProviderImpl.kt
@@ -36,7 +36,7 @@ class PredefinedMoodTypeProviderImpl(
 
     private fun create(
         @StringRes nameResId: Int,
-        iconId: Long,
+        iconId: Int,
         positivePercentage: Int
     ) = MoodTypePredefined(
         name = context.getString(nameResId),
